Feature ETP-3778: address SIF PR follow-up
