+++ artifacts/sales-invoice/custom/__tests__/SendToSifButton.test.js
@@ -8,81 +8,7 @@ const __dirname = dirname(fileURLToPath(import.meta.url));
 const src = readFileSync(join(__dirname, '..', 'SendToSifButton.jsx'), 'utf8');
 
 describe('SendToSifButton', () => {
-
   it('re-exports the shared SendToSifButton implementation', () => {
     assert.match(src, /export\s*\{\s*default\s*\}\s*from\s*['"]@\/windows\/custom\/shared\/SendToSifButton\.jsx['"]/);
-  });
-
-  it('exports a default function component named SendToSifButton', () => {
-    assert.match(src, /export\s*\{\s*default\s*\}/);
-  });
-
-  it('accepts data, recordId, token, apiBaseUrl, and status props', () => {
-    assert.match(src, /shared\/SendToSifButton/);
-  });
-
-  it('returns null when status is not CO', () => {
-    assert.match(src, /SendToSifButton/);
-  });
-
-  it('uses useFiscalConfig hook', () => {
-    assert.match(src, /shared\/SendToSifButton/);
-  });
-
-  it('imports useFiscalConfig from fiscal-config path', () => {
-    assert.match(src, /shared\/SendToSifButton/);
-  });
-
-  it('defines VISIBLE_PROFILES set with sii, tbai, and sii+tbai profiles', () => {
-    assert.match(src, /shared\/SendToSifButton/);
-    assert.match(src, /SendToSifButton/);
-  });
-
-  it('returns null when profile is not in VISIBLE_PROFILES', () => {
-    assert.match(src, /shared\/SendToSifButton/);
-  });
-
-  it('calls SII endpoint Em_aeatsii_send', () => {
-    assert.match(src, /shared\/SendToSifButton/);
-  });
-
-  it('calls TBAI endpoint Em_Tbai_Xmlgenerator', () => {
-    assert.match(src, /shared\/SendToSifButton/);
-  });
-
-  it('uses three modal phases: confirm, sending, results', () => {
-    assert.match(src, /shared\/SendToSifButton/);
-  });
-
-  it('sets phase to sending before fetch calls', () => {
-    assert.match(src, /shared\/SendToSifButton/);
-  });
-
-  it('sets phase to results after fetch calls', () => {
-    assert.match(src, /shared\/SendToSifButton/);
-  });
-
-  it('handles SII and TBAI independently (separate try/catch)', () => {
-    assert.match(src, /shared\/SendToSifButton/);
-  });
-
-  it('uses sendToSif i18n key for button label', () => {
-    assert.match(src, /shared\/SendToSifButton/);
-  });
-
-  it('uses sendToSifTitle i18n key for modal title', () => {
-    assert.match(src, /shared\/SendToSifButton/);
-  });
-
-  it('uses sendToSifSending i18n key during send phase', () => {
-    assert.match(src, /shared\/SendToSifButton/);
-  });
-
-  it('shows success and error result rows using i18n keys', () => {
-    assert.match(src, /shared\/SendToSifButton/);
-  });
-
-  it('derives orgId from selectedOrg via useAuth', () => {
-    assert.match(src, /shared\/SendToSifButton/);
   });
 });
