+++ tools/app-shell/src/locales/en_US.json
@@ -18145,7 +18145,45 @@
     "sendToSifSuccessSii": "Sent to SII successfully.",
     "sendToSifSuccessTbai": "Sent to TBAI successfully.",
     "sendToSifErrorSii": "Error sending to SII.",
-    "sendToSifErrorTbai": "Error sending to TBAI."
+    "sendToSifErrorTbai": "Error sending to TBAI.",
+    "sifDataTabs.sectionTitle": "SIF",
+    "sifDataTabs.tab.sii": "SII",
+    "sifDataTabs.tab.tbai": "TBAI",
+    "sifDataTabs.tab.verifactu": "Verifactu",
+    "sifDataTabs.errorSavingField": "Error saving field",
+    "sifDataTabs.field.operationDate": "Operation date",
+    "sifDataTabs.field.invoiceType": "Invoice type",
+    "sifDataTabs.field.masterDescription": "Master description",
+    "sifDataTabs.field.siiDescription": "SII description",
+    "sifDataTabs.field.exemptionCause": "Exemption cause",
+    "sifDataTabs.field.authorization": "Authorization",
+    "sifDataTabs.field.siiYear": "SII year",
+    "sifDataTabs.field.siiPeriod": "SII period",
+    "sifDataTabs.field.chainSequence": "Chain sequence",
+    "sifDataTabs.field.invoiceSeries": "Invoice series",
+    "sifDataTabs.field.invoiceSequence": "Invoice sequence",
+    "sifDataTabs.field.rfGenerationDate": "RF generation date",
+    "sifDataTabs.field.csv": "CSV",
+    "sifDataTabs.field.hash": "Hash",
+    "sifDataTabs.field.qrUrl": "QR URL",
+    "sifDataTabs.field.issueDetail": "Issue detail",
+    "sifDataTabs.status.sii.correct": "Correct",
+    "sifDataTabs.status.sii.acceptedWithErrors": "Accepted with errors",
+    "sifDataTabs.status.sii.incorrect": "Incorrect",
+    "sifDataTabs.status.sii.sendError": "Send error",
+    "sifDataTabs.status.sii.pending": "Pending",
+    "sifDataTabs.status.sii.cancelled": "Cancelled",
+    "sifDataTabs.status.sii.dropped": "Dropped",
+    "sifDataTabs.status.sii.notRegistrable": "Not registrable",
+    "sifDataTabs.status.sii.notSent": "Not sent",
+    "sifDataTabs.status.tbai.sent": "Sent to TBAI",
+    "sifDataTabs.status.tbai.notSent": "Not sent",
+    "sifDataTabs.status.verifactu.accepted": "Accepted",
+    "sifDataTabs.status.verifactu.acceptedWithErrors": "Accepted with errors",
+    "sifDataTabs.status.verifactu.invalid": "Invalid",
+    "sifDataTabs.status.verifactu.rejected": "Rejected",
+    "sifDataTabs.status.verifactu.pending": "Pending",
+    "sifDataTabs.status.verifactu.notSent": "Not sent"
   },
   "menus": {
     "Currency": {
