+++ tools/app-shell/src/locales/es_ES.json
@@ -18390,7 +18390,45 @@
     "sendToSifSuccessSii": "Enviado al SII correctamente.",
     "sendToSifSuccessTbai": "Enviado a TBAI correctamente.",
     "sendToSifErrorSii": "Error al enviar al SII.",
-    "sendToSifErrorTbai": "Error al enviar a TBAI."
+    "sendToSifErrorTbai": "Error al enviar a TBAI.",
+    "sifDataTabs.sectionTitle": "SIF",
+    "sifDataTabs.tab.sii": "SII",
+    "sifDataTabs.tab.tbai": "TBAI",
+    "sifDataTabs.tab.verifactu": "Verifactu",
+    "sifDataTabs.errorSavingField": "Error al guardar el campo",
+    "sifDataTabs.field.operationDate": "Fecha operación",
+    "sifDataTabs.field.invoiceType": "Tipo factura",
+    "sifDataTabs.field.masterDescription": "Descripción maestra",
+    "sifDataTabs.field.siiDescription": "Descripción SII",
+    "sifDataTabs.field.exemptionCause": "Causa exención",
+    "sifDataTabs.field.authorization": "Autorización",
+    "sifDataTabs.field.siiYear": "Ejercicio SII",
+    "sifDataTabs.field.siiPeriod": "Periodo SII",
+    "sifDataTabs.field.chainSequence": "Secuencia encadenamiento",
+    "sifDataTabs.field.invoiceSeries": "Serie factura",
+    "sifDataTabs.field.invoiceSequence": "Secuencia factura",
+    "sifDataTabs.field.rfGenerationDate": "Fecha generación RF",
+    "sifDataTabs.field.csv": "CSV",
+    "sifDataTabs.field.hash": "Hash",
+    "sifDataTabs.field.qrUrl": "QR URL",
+    "sifDataTabs.field.issueDetail": "Detalle incidencia",
+    "sifDataTabs.status.sii.correct": "Correcto",
+    "sifDataTabs.status.sii.acceptedWithErrors": "Aceptado con errores",
+    "sifDataTabs.status.sii.incorrect": "Incorrecto",
+    "sifDataTabs.status.sii.sendError": "Error de envío",
+    "sifDataTabs.status.sii.pending": "Pendiente",
+    "sifDataTabs.status.sii.cancelled": "Anulada",
+    "sifDataTabs.status.sii.dropped": "Dada de baja",
+    "sifDataTabs.status.sii.notRegistrable": "No registrable",
+    "sifDataTabs.status.sii.notSent": "No enviada",
+    "sifDataTabs.status.tbai.sent": "Enviada a TBAI",
+    "sifDataTabs.status.tbai.notSent": "No enviada",
+    "sifDataTabs.status.verifactu.accepted": "Aceptada",
+    "sifDataTabs.status.verifactu.acceptedWithErrors": "Aceptada con errores",
+    "sifDataTabs.status.verifactu.invalid": "Inválida",
+    "sifDataTabs.status.verifactu.rejected": "Rechazada",
+    "sifDataTabs.status.verifactu.pending": "Pendiente",
+    "sifDataTabs.status.verifactu.notSent": "No enviada"
   },
   "menus": {
     "Tax": {
