+++ tools/app-shell/src/windows/custom/purchase-invoice/PurchaseInvoiceBottomPanel.jsx
@@ -1,6 +1,6 @@
 import { useUI } from '@/i18n';
 import RelatedDocuments from './RelatedDocuments.jsx';
-import SifDataTabs from '@generated/sales-invoice/custom/SifDataTabs';
+import SifDataTabs from '@/windows/custom/shared/SifDataTabs.jsx';
 import { formatCurrency } from '@/lib/formatCurrency';
 import DocumentTotalsPanel from '@/components/contract-ui/DocumentTotalsPanel.jsx';
